Add devicePixelRatio to PaintWorkletGlobalScope
In this CL, we add the devicePixelRatio property to the
PaintWorkletGlobalScope.idl. This property could be useful for
developers to draw perfect graphics on high DPI devices.

A test is added to ensure that this property has the same value as the
window.devicePixelRatio.

Bug: 728599
Change-Id: Ia6179164989d4f564df06a87f1399adff19a0599
Reviewed-on: https://chromium-review.googlesource.com/602367
WPT-Export-Revision: e073640d554254ba342986906518b2e3b68765b4